Use token ranges

+++ Api/src/UrlShortener.Core/Base62EncodingExtensions.cs
@@ -4,7 +4,7 @@ public static class Base62EncodingExtensions
 {
     private const string Alphanumeric = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
 
-    public static string EncodeToBase62(this int number)
+    public static string EncodeToBase62(this long number)
     {
         if (number == 0) return Alphanumeric[0].ToString();
 
@@ -12,7 +12,7 @@ public static class Base62EncodingExtensions
 
         while (number > 0)
         {
-            result.Push(Alphanumeric[number % Alphanumeric.Length]);
+            result.Push(Alphanumeric[(int)number % Alphanumeric.Length]);
             number /= Alphanumeric.Length;
         }
